Relocation injectors package
