+++ bcs-app/backend/kube_core/toolkit/injectors/dpath.py
